Multiple changes
B2 worker scripts - adds in sections to save the regression coefficient outputs so we can see if the model isn't converging or any other weird behavior if the results end up being the crazy huge numbers.

A2 F2T worker - comments out the section that only filters on primary cause = 1, we want to test to see if possibly removing this filtration solves the issue of the crazy huge numbers.